qcacld-2.0: Implement attribute for SET WIFI config for ifindex
Implement QCA_WLAN_VENDOR_ATTR_CONFIG_NON_AGG_RETRY and
QCA_WLAN_VENDOR_ATTR_CONFIG_AGG_RETRY for SET wifi config as per
index as well.

CRs-Fixed: 1040532

+++ CORE/SME/inc/sme_Api.h
@@ -4594,9 +4594,9 @@ eHalStatus sme_update_access_policy_vendor_ie(tHalHandle hal,
 eHalStatus sme_update_tx_fail_cnt_threshold(tHalHandle hal_handle,
 		uint8_t session_id, uint32_t tx_fail_count);
 eHalStatus sme_update_short_retry_limit_threshold(tHalHandle hal_handle,
-		uint8_t session_id, uint32_t short_limit_count_th);
+		uint8_t session_id, uint8_t short_limit_count_th);
 eHalStatus sme_update_long_retry_limit_threshold(tHalHandle hal_handle,
-		uint8_t session_id, uint32_t long_limit_count_th);
+		uint8_t session_id, uint8_t long_limit_count_th);
 eHalStatus sme_update_sta_inactivity_timeout(tHalHandle hal_handle,
 		uint8_t session_id, uint32_t sta_inactivity_timeout);
 #endif //#if !defined( __SME_API_H )
